refactor: rename @adcio.js -> @adcio
