+++ demo/react/vite.config.ts
@@ -5,7 +5,7 @@ import tsconfigPaths from "vite-tsconfig-paths";
 export default defineConfig({
   build: {
     rollupOptions: {
-      external: ["@adcio.js/core"],
+      external: ["@adcio/core"],
     }
   },
   plugins: [react(), tsconfigPaths()],
