+++ packages/api/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "@adcio.js/api",
+  "name": "@adcio/api",
   "private": true
 }
