+++ packages/core/src/index.ts
@@ -12,10 +12,10 @@ export type {
 } from "./adcio.interface";
 
 // FrontAPI
-export type { ClientAPI } from "@adcio.js/lib/client-api";
-export { Cafe24API } from "@adcio.js/lib/client-api";
+export type { ClientAPI } from "@adcio/lib/client-api";
+export { Cafe24API } from "@adcio/lib/client-api";
 
-import { Cafe24API } from "@adcio.js/lib/client-api";
+import { Cafe24API } from "@adcio/lib/client-api";
 export const clientApi = {
   cafe24: new Cafe24API(),
 };
@@ -27,4 +27,4 @@ export {
   getMeta,
   createElement,
   createNestedElement,
-} from "@adcio.js/lib/utils";
+} from "@adcio/lib/utils";
