+++ packages/lib/placement/placement.interface.ts
@@ -2,7 +2,7 @@ import {
   BannerSuggestionRequestDto,
   ProductSuggestionRequestDto,
   SuggestionRequestDto,
-} from "@adcio.js/api/controller/v1";
+} from "@adcio/api/controller/v1";
 import { AdcioCore } from "../core";
 
 export type AdcioPlacementParams = {
